refactor: rename controlled unitary module and imports
Agent-Logs-Url: https://github.com/microsoft/qdk-chemistry/sessions/af236b7a-4f73-4db4-8b82-779d69d8618e
